Assignment1/requirements.txt
fastapi
flask
mlflow
numpy
pandas
prometheus_client
pyarrow
PyYAML
pytest
pytest-cov
requests
scikit-learn
scipy
uvicorn
boto3
botocore
pyspark
findspark
evidently
s3fs
great_expectations
